FC/TourismDigitalFC/requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
pmdarima>=2.0
statsmodels>=0.14
prophet>=1.1
scikit-learn>=1.3
fastapi>=0.104
uvicorn>=0.24
pydantic>=2.0
python-multipart>=0.0.6
motor>=3.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
passlib>=1.7.4
bcrypt<4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
certifi>=2020.12.5